ersdf

+++ facefusion/processors/frame/modules/face_swapper.py
@@ -268,7 +268,7 @@ def process_frame(source_face : Face, reference_faces : FaceSet, temp_frame : Fr
 			for similar_face in similar_faces:
 				temp_frame = swap_face(source_face, similar_face, temp_frame)
 		else:
-			if(temp_frame_path != ""):
+			if(temp_frame_path != "" and facefusion.globals.skip_audio):
 				os.remove(temp_frame_path)
 				return False
 	if 'one' in facefusion.globals.face_selector_mode:
